// interfaces/menu.interface.ts
import {JSX} from "react";
import {TopLevelCategory} from "@/interfaces/page.interface";

export interface PageItem {
    alias: string;
    title: string;
    _id: string;
    category: string;
}

export interface MenuItem {
    _id: {
        secondCategory: string
    };
    isOpened?: boolean;
    pages: PageItem[];
}

export interface FirstLeveMenuItem {
    route: string;
    name: string;
    icon: JSX.Element;
    id: TopLevelCategory;
}